4 character controllers
There are 4 character controllers in this project. With increased number to the characterController the features gets incresed too.
CharacterController1 - Is a simple character controller which allows to control the player using arrow keys
CharacterController2- gives the ability to jump
CharacterController3 - Rotates the player in the horizontal axis with the movement of mouse
CharacterController4- Rotates the camera in vertical direction.